+++ src/Markdown/demos/streamingProfiler.tsx
@@ -1,5 +1,5 @@
 import StreamingPlayground from './StreamingPlayground';
 
 export default () => {
-  return <StreamingPlayground />;
+  return <StreamingPlayground defaultShowProfiler />;
 };